+++ example/prometheus/main.go
@@ -76,11 +76,11 @@ func main() {
 		(*observerLock).RUnlock()
 		result.Observe(value, labels...)
 	}
-	_ = metric.Must(meter).NewFloat64ValueObserver("ex.com.one", cb,
-		metric.WithDescription("A ValueObserver set to 1.0"),
+	_ = metric.Must(meter).NewFloat64GaugeObserver("ex.com.one", cb,
+		metric.WithDescription("A GaugeObserver set to 1.0"),
 	)
 
-	valuerecorder := metric.Must(meter).NewFloat64ValueRecorder("ex.com.two")
+	histogram := metric.Must(meter).NewFloat64Histogram("ex.com.two")
 	counter := metric.Must(meter).NewFloat64Counter("ex.com.three")
 
 	commonLabels := []attribute.KeyValue{lemonsKey.Int(10), attribute.String("A", "1"), attribute.String("B", "2"), attribute.String("C", "3")}
@@ -95,7 +95,7 @@ func main() {
 	meter.RecordBatch(
 		ctx,
 		commonLabels,
-		valuerecorder.Measurement(2.0),
+		histogram.Measurement(2.0),
 		counter.Measurement(12.0),
 	)
 
@@ -108,7 +108,7 @@ func main() {
 	meter.RecordBatch(
 		ctx,
 		notSoCommonLabels,
-		valuerecorder.Measurement(2.0),
+		histogram.Measurement(2.0),
 		counter.Measurement(22.0),
 	)
 
@@ -121,7 +121,7 @@ func main() {
 	meter.RecordBatch(
 		ctx,
 		commonLabels,
-		valuerecorder.Measurement(12.0),
+		histogram.Measurement(12.0),
 		counter.Measurement(13.0),
 	)