+++ internal/metric/global/meter_test.go
@@ -58,21 +58,21 @@ func TestDirect(t *testing.T) {
 	counter.Add(ctx, 1, labels1...)
 	counter.Add(ctx, 1, labels1...)
 
-	valuerecorder := Must(meter1).NewFloat64ValueRecorder("test.valuerecorder")
-	valuerecorder.Record(ctx, 1, labels1...)
-	valuerecorder.Record(ctx, 2, labels1...)
+	histogram := Must(meter1).NewFloat64Histogram("test.histogram")
+	histogram.Record(ctx, 1, labels1...)
+	histogram.Record(ctx, 2, labels1...)
 
-	_ = Must(meter1).NewFloat64ValueObserver("test.valueobserver.float", func(_ context.Context, result metric.Float64ObserverResult) {
+	_ = Must(meter1).NewFloat64GaugeObserver("test.gauge.float", func(_ context.Context, result metric.Float64ObserverResult) {
 		result.Observe(1., labels1...)
 		result.Observe(2., labels2...)
 	})
 
-	_ = Must(meter1).NewInt64ValueObserver("test.valueobserver.int", func(_ context.Context, result metric.Int64ObserverResult) {
+	_ = Must(meter1).NewInt64GaugeObserver("test.gauge.int", func(_ context.Context, result metric.Int64ObserverResult) {
 		result.Observe(1, labels1...)
 		result.Observe(2, labels2...)
 	})
 
-	second := Must(meter2).NewFloat64ValueRecorder("test.second")
+	second := Must(meter2).NewFloat64Histogram("test.second")
 	second.Record(ctx, 1, labels3...)
 	second.Record(ctx, 2, labels3...)
 
@@ -80,7 +80,7 @@ func TestDirect(t *testing.T) {
 	metricglobal.SetMeterProvider(provider)
 
 	counter.Add(ctx, 1, labels1...)
-	valuerecorder.Record(ctx, 3, labels1...)
+	histogram.Record(ctx, 3, labels1...)
 	second.Record(ctx, 3, labels3...)
 
 	provider.RunAsyncInstruments()
@@ -96,7 +96,7 @@ func TestDirect(t *testing.T) {
 				Number:  asInt(1),
 			},
 			{
-				Name:    "test.valuerecorder",
+				Name:    "test.histogram",
 				Library: library1,
 				Labels:  metrictest.LabelsToMap(labels1...),
 				Number:  asFloat(3),
@@ -108,25 +108,25 @@ func TestDirect(t *testing.T) {
 				Number:  asFloat(3),
 			},
 			{
-				Name:    "test.valueobserver.float",
+				Name:    "test.gauge.float",
 				Library: library1,
 				Labels:  metrictest.LabelsToMap(labels1...),
 				Number:  asFloat(1),
 			},
 			{
-				Name:    "test.valueobserver.float",
+				Name:    "test.gauge.float",
 				Library: library1,
 				Labels:  metrictest.LabelsToMap(labels2...),
 				Number:  asFloat(2),
 			},
 			{
-				Name:    "test.valueobserver.int",
+				Name:    "test.gauge.int",
 				Library: library1,
 				Labels:  metrictest.LabelsToMap(labels1...),
 				Number:  asInt(1),
 			},
 			{
-				Name:    "test.valueobserver.int",
+				Name:    "test.gauge.int",
 				Library: library1,
 				Labels:  metrictest.LabelsToMap(labels2...),
 				Number:  asInt(2),
@@ -154,8 +154,8 @@ func TestBound(t *testing.T) {
 	boundC.Add(ctx, 1)
 	boundC.Add(ctx, 1)
 
-	valuerecorder := Must(glob).NewInt64ValueRecorder("test.valuerecorder")
-	boundM := valuerecorder.Bind(labels1...)
+	histogram := Must(glob).NewInt64Histogram("test.histogram")
+	boundM := histogram.Bind(labels1...)
 	boundM.Record(ctx, 1)
 	boundM.Record(ctx, 2)
 
@@ -180,7 +180,7 @@ func TestBound(t *testing.T) {
 				Number:  asFloat(1),
 			},
 			{
-				Name:    "test.valuerecorder",
+				Name:    "test.histogram",
 				Library: library,
 				Labels:  metrictest.LabelsToMap(labels1...),
 				Number:  asInt(3),
@@ -202,8 +202,8 @@ func TestUnbind(t *testing.T) {
 	counter := Must(glob).NewFloat64Counter("test.counter")
 	boundC := counter.Bind(labels1...)
 
-	valuerecorder := Must(glob).NewInt64ValueRecorder("test.valuerecorder")
-	boundM := valuerecorder.Bind(labels1...)
+	histogram := Must(glob).NewInt64Histogram("test.histogram")
+	boundM := histogram.Bind(labels1...)
 
 	boundC.Unbind()
 	boundM.Unbind()
@@ -282,12 +282,12 @@ func TestImplementationIndirection(t *testing.T) {
 	require.False(t, ok)
 
 	// Async: no SDK yet
-	valueobserver := Must(meter1).NewFloat64ValueObserver(
-		"interface.valueobserver",
+	gauge := Must(meter1).NewFloat64GaugeObserver(
+		"interface.gauge",
 		func(_ context.Context, result metric.Float64ObserverResult) {},
 	)
 
-	ival = valueobserver.AsyncImpl().Implementation()
+	ival = gauge.AsyncImpl().Implementation()
 	require.NotNil(t, ival)
 
 	_, ok = ival.(*metrictest.Async)
@@ -307,7 +307,7 @@ func TestImplementationIndirection(t *testing.T) {
 	require.True(t, ok)
 
 	// Async
-	ival = valueobserver.AsyncImpl().Implementation()
+	ival = gauge.AsyncImpl().Implementation()
 	require.NotNil(t, ival)
 
 	_, ok = ival.(*metrictest.Async)